requirements.txt
tensorflow-gpu
ray=0.6.0
jupyter
requests
pillow
psutil
setproctitle

